// src/types/theme.ts
export interface ThemeFonts {
    text: string;
    h1: string;
    h2: string;
    h3: string;
    h4: string;
    h5: string;
    p: string;
    thin: string;
    extralight: string;
    light: string;
    normal: string;
    medium: string;
    bold: string;
    semibold: string;
    extrabold: string;
    black: string;
}